+++ src/AgentFrameworkToolkit.AzureOpenAI/AzureOpenAIAgentFactory.cs
@@ -1,10 +1,12 @@
 ﻿using AgentFrameworkToolkit.OpenAI;
 using Azure.AI.OpenAI;
+using Azure.Core;
 using JetBrains.Annotations;
 using Microsoft.Agents.AI;
 using Microsoft.Extensions.AI;
 using OpenAI;
-using Azure.Core;
+using OpenAI.Chat;
+using OpenAI.Responses;
 
 #pragma warning disable OPENAI001
 
@@ -66,8 +68,9 @@ public class AzureOpenAIAgentFactory
     /// <returns>An Agent</returns>
     public AzureOpenAIAgent CreateAgent(string model, string? instructions = null, string? name = null, AITool[]? tools = null)
     {
-        return CreateAgent(new OpenAIAgentOptionsForChatClientWithoutReasoning
+        return CreateAgent(new AgentOptions
         {
+            ClientType = ClientType.ChatClient,
             Model = model,
             Name = name,
             Instructions = instructions,
@@ -80,6 +83,44 @@ public class AzureOpenAIAgentFactory
     /// </summary>
     /// <param name="options">Options for the agent</param>
     /// <returns>The Agent</returns>
+    public AzureOpenAIAgent CreateAgent(AgentOptions options)
+    {
+        OpenAIClient client = _connection.GetClient(options.RawHttpCallDetails);
+
+        ChatClientAgentOptions chatClientAgentOptions = CreateChatClientAgentOptions(options);
+
+        ChatClientAgent innerAgent;
+        switch (options.ClientType)
+        {
+            case ClientType.ChatClient:
+                innerAgent = client
+                    .GetChatClient(options.Model)
+                    .CreateAIAgent(chatClientAgentOptions);
+                break;
+            case ClientType.ResponsesApi:
+                innerAgent = client
+                    .GetOpenAIResponseClient(options.Model)
+                    .CreateAIAgent(chatClientAgentOptions);
+                break;
+            default:
+                throw new ArgumentOutOfRangeException();
+        }
+
+        // ReSharper disable once ConvertIfStatementToReturnStatement
+        if (options.RawToolCallDetails != null)
+        {
+            return new AzureOpenAIAgent(options.ApplyMiddleware(innerAgent));
+        }
+
+        return new AzureOpenAIAgent(innerAgent);
+    }
+
+    /// <summary>
+    /// Create a new Agent
+    /// </summary>
+    /// <param name="options">Options for the agent</param>
+    /// <returns>The Agent</returns>
+    [Obsolete("Use 'AgentOptions' variant instead (This method will be remove 1st of January 2026)")]
     public AzureOpenAIAgent CreateAgent(OpenAIAgentOptionsForResponseApiWithoutReasoning options)
     {
         AzureOpenAIClient client = _connection.GetClient(options.RawHttpCallDetails);
@@ -104,6 +145,7 @@ public class AzureOpenAIAgentFactory
     /// </summary>
     /// <param name="options">Options for the agent</param>
     /// <returns>The Agent</returns>
+    [Obsolete("Use 'AgentOptions' variant instead (This method will be remove 1st of January 2026)")]
     public AzureOpenAIAgent CreateAgent(OpenAIAgentOptionsForResponseApiWithReasoning options)
     {
         AzureOpenAIClient client = _connection.GetClient(options.RawHttpCallDetails);
@@ -128,6 +170,7 @@ public class AzureOpenAIAgentFactory
     /// </summary>
     /// <param name="options">Options for the agent</param>
     /// <returns>The Agent</returns>
+    [Obsolete("Use 'AgentOptions' variant instead (This method will be remove 1st of January 2026)")]
     public AzureOpenAIAgent CreateAgent(OpenAIAgentOptionsForChatClientWithoutReasoning options)
     {
         AzureOpenAIClient client = _connection.GetClient(options.RawHttpCallDetails);
@@ -156,6 +199,7 @@ public class AzureOpenAIAgentFactory
     /// </summary>
     /// <param name="options">Options for the agent</param>
     /// <returns>The Agent</returns>
+    [Obsolete("Use 'AgentOptions' variant instead (This method will be remove 1st of January 2026)")]
     public AzureOpenAIAgent CreateAgent(OpenAIAgentOptionsForChatClientWithReasoning options)
     {
         AzureOpenAIClient client = _connection.GetClient(options.RawHttpCallDetails);
@@ -179,6 +223,7 @@ public class AzureOpenAIAgentFactory
         return new AzureOpenAIAgent(innerAgent);
     }
 
+    [Obsolete("Use 'AgentOptions' variant instead (This method will be remove 1st of January 2026)")]
     private ChatClientAgentOptions CreateChatClientAgentOptions(OpenAIAgentOptions options, OpenAIAgentOptionsForChatClientWithoutReasoning? chatClientWithoutReasoning, OpenAIAgentOptionsForResponseApiWithoutReasoning? responseWithoutReasoning, OpenAIAgentOptionsForResponseApiWithReasoning? responsesApiReasoningOptions, OpenAIAgentOptionsForChatClientWithReasoning? chatClientReasoningOptions)
     {
         bool anyOptionsSet = false;
@@ -228,6 +273,105 @@ public class AzureOpenAIAgentFactory
             chatOptions = chatOptions.WithOpenAIChatClientReasoning(chatClientReasoningOptions.ReasoningEffort);
         }
 
+
+        ChatClientAgentOptions chatClientAgentOptions = new()
+        {
+            Name = options.Name,
+            Description = options.Description,
+            Id = options.Id,
+        };
+        if (anyOptionsSet)
+        {
+            chatClientAgentOptions.ChatOptions = chatOptions;
+        }
+
+        options.AdditionalChatClientAgentOptions?.Invoke(chatClientAgentOptions);
+
+        return chatClientAgentOptions;
+    }
+
+    private ChatClientAgentOptions CreateChatClientAgentOptions(AgentOptions options)
+    {
+        bool anyOptionsSet = false;
+        ChatOptions chatOptions = new();
+        if (options.Tools != null)
+        {
+            anyOptionsSet = true;
+            chatOptions.Tools = options.Tools;
+        }
+
+        if (options.MaxOutputTokens.HasValue)
+        {
+            anyOptionsSet = true;
+            chatOptions.MaxOutputTokens = options.MaxOutputTokens.Value;
+        }
+
+        if (options.Temperature != null && !OpenAIChatModels.ReasoningModels.Contains(options.Model))
+        {
+            anyOptionsSet = true;
+            chatOptions.Temperature = options.Temperature;
+        }
+
+        if (!string.IsNullOrWhiteSpace(options.Instructions))
+        {
+            anyOptionsSet = true;
+            chatOptions.Instructions = options.Instructions;
+        }
+
+        string? reasoningEffortAsString = null;
+        switch (options.ReasoningEffort)
+        {
+            case OpenAIReasoningEffort.None:
+                reasoningEffortAsString = "none";
+                break;
+            case OpenAIReasoningEffort.Minimal:
+                reasoningEffortAsString = "minimal";
+                break;
+            case OpenAIReasoningEffort.Low:
+                reasoningEffortAsString = "low";
+                break;
+            case OpenAIReasoningEffort.Medium:
+                reasoningEffortAsString = "medium";
+                break;
+            case OpenAIReasoningEffort.High:
+                reasoningEffortAsString = "high";
+                break;
+            case OpenAIReasoningEffort.ExtraHigh:
+                reasoningEffortAsString = "xhigh";
+                break;
+        }
+
+        if (!string.IsNullOrWhiteSpace(reasoningEffortAsString) && !OpenAIChatModels.NonReasoningModels.Contains(options.Model))
+        {
+            anyOptionsSet = true;
+
+            switch (options.ClientType)
+            {
+                case ClientType.ChatClient:
+                    chatOptions = chatOptions.WithOpenAIChatClientReasoning(new ChatReasoningEffortLevel(reasoningEffortAsString));
+                    break;
+                case ClientType.ResponsesApi:
+                    switch (options.ReasoningSummaryVerbosity)
+                    {
+                        case OpenAIReasoningSummaryVerbosity.Auto:
+                            chatOptions = chatOptions.WithOpenAIResponsesApiReasoning(new ResponseReasoningEffortLevel(reasoningEffortAsString), ResponseReasoningSummaryVerbosity.Auto);
+                            break;
+                        case OpenAIReasoningSummaryVerbosity.Concise:
+                            chatOptions = chatOptions.WithOpenAIResponsesApiReasoning(new ResponseReasoningEffortLevel(reasoningEffortAsString), ResponseReasoningSummaryVerbosity.Concise);
+                            break;
+                        case OpenAIReasoningSummaryVerbosity.Detailed:
+                            chatOptions = chatOptions.WithOpenAIResponsesApiReasoning(new ResponseReasoningEffortLevel(reasoningEffortAsString), ResponseReasoningSummaryVerbosity.Detailed);
+                            break;
+                        case null:
+                            chatOptions = chatOptions.WithOpenAIResponsesApiReasoning(new ResponseReasoningEffortLevel(reasoningEffortAsString));
+                            break;
+                    }
+
+                    break;
+                default:
+                    throw new ArgumentOutOfRangeException();
+            }
+        }
 
         ChatClientAgentOptions chatClientAgentOptions = new()
         {