+++ src/AgentFrameworkToolkit.OpenAI/OpenAIReasoningSummaryVerbosity.cs
@@ -1,0 +1,22 @@
+﻿namespace AgentFrameworkToolkit.OpenAI;
+
+/// <summary>
+/// The Verbosity of the Reasoning Summary [Responses API Only]
+/// </summary>
+public enum OpenAIReasoningSummaryVerbosity
+{
+    /// <summary>
+    /// Auto
+    /// </summary>
+    Auto,
+
+    /// <summary>
+    /// Concise
+    /// </summary>
+    Concise,
+
+    /// <summary>
+    /// Detailed
+    /// </summary>
+    Detailed
+}